third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/scoped-registry/ShadowRoot-init-registry.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] ShadowRoot.registry is null if not explicitly specified
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] Attach the global registry to a shadow root
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] Attach a non-global registry to a shadow root
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Attach the same registry to multiple shadow roots
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Attaching registry to shadow root can only be done during initialization
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
